+++ src/FluentDbTools/Example/Example.FluentDbTools.MigrationWithVersionTable/ExampleVersionTable.cs
@@ -1,0 +1,10 @@
+﻿using System;
+namespace Example.FluentDbTools.MigrationWithVersionTable
+{
+    public class ExampleVersionTable
+    {
+        public string Version { get; set; }
+        public DateTimeOffset AppliedOn { get; set; }
+        public string Description { get; set; }
+    }
+}